create code
The beginnings of the code, the in and out and defining of the pins (not built yet so no pins included). the lcd display needs to be added as well as the max number of people. we hope to add a keypad to disable the buzzer that will sound once the maximum number of people within the smart room has been exceeded.